feat: Add DSL::Animate
